pyproject.toml
[project]
name = "recipes-api"
version = "0.1.0"
description = ""

authors = [
    { name = "hyperskill", email = "hyper@hyperskill.org" }
]

readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django>=5.2,<6.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "bandit (>=1.8.3,<2.0.0)",
    "flake8-json (>=24.4.0,<25.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-json-report = ">=1.5.0,<2.0.0"
coverage = "^7.8.0"
flake8 = "^7.2.0"
flake8-json = ">=24.4.0,<25.0.0"
jq = ">=1.8.0,<2.0.0"


[tool.coverage.run]
relative_files = true
